Move replay buffer
